refactor: 가게 밀도 검색 모듈화
